refactor: break util into multiple submodules
