+++ weasel/cli/assets.py
@@ -10,7 +10,7 @@ from wasabi import msg
 
 from .._util import SimpleFrozenDict, download_file, get_checksum, get_git_version
 from .._util import git_checkout, load_project_config, parse_config_overrides
-from ..util import ensure_path, working_dir
+from ..utils import ensure_path, working_dir
 from .main import PROJECT_FILE, Arg, Opt, app
 
 # Whether assets are extra if `extra` is not set.
