+++ weasel/cli/document.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from wasabi import MarkdownRenderer, msg
 
 from .._util import load_project_config
-from ..util import working_dir
+from ..utils import working_dir
 from .main import PROJECT_FILE, Arg, Opt, app
 
 DOCS_URL = "https://spacy.io"
