+++ weasel/utils/logging.py
@@ -1,0 +1,8 @@
+import logging
+
+logger = logging.getLogger("weasel")
+logger_stream_handler = logging.StreamHandler()
+logger_stream_handler.setFormatter(
+    logging.Formatter("[%(asctime)s] [%(levelname)s] %(message)s")
+)
+logger.addHandler(logger_stream_handler)
